# Add broken pair reader