--- a/lib/read_parsers.hh
+++ b/lib/read_parsers.hh
@@ -109,7 +109,8 @@ class IParser
 {
 public:
     enum {
-        PAIR_MODE_IGNORE_UNPAIRED = 1,
+        PAIR_MODE_ALLOW_UNPAIRED = 0,
+        PAIR_MODE_IGNORE_UNPAIRED,
         PAIR_MODE_ERROR_ON_UNPAIRED
     };