--- a/scripts/extract-paired-reads.py
+++ b/scripts/extract-paired-reads.py
@@ -50,7 +50,6 @@ import sys
 import os.path
 import textwrap
 
-from khmer import __version__
 from khmer import ReadParser
 from khmer.kfile import check_input_files, check_space
 from khmer.khmer_args import sanitize_help, KhmerArgumentParser
@@ -154,9 +153,11 @@ def main():
     n_pe = 0
     n_se = 0
 
-    screed_iter = ReadParser(infile)
-    #for index, is_pair, read1, read2 in broken_paired_reader(screed_iter):
-    for index, (read1, read2) in enumerate(screed_iter.iter_read_pairs(0)):
+    reads_iter = ReadParser(infile)
+    # python version
+    # for index, is_pair, read1, read2 in broken_paired_reader(reads_iter):
+    # c++ version
+    for index, (read1, read2) in enumerate(reads_iter.iter_read_pairs(0)):
         is_pair = read2 is not None
         if index % 100000 == 0 and index > 0:
             print('...', index, file=sys.stderr)