@betatim 2016-11-21
Totally not optimised c++, benchmarking `time scripts/extract-paired-reads.py ecoli_ref-5m.fastq.gz -s /dev/null -p /dev/null`:

* `master`: 71.77s user 0.53s system 99% cpu 1:12.65 total
* this PR: 56.67s user 0.89s system 98% cpu 58.397 total

😕