@ctb 2016-11-21
Do you think that's because the Python side is slowing stuff down a lot?