@betatim 2016-11-22
There are a few things that I am confused about/need some educating.

Why the `PyObject * obj = khmer_ReadPairIterator_Type.tp_alloc(&khmer_ReadPairIterator_Type, 1);` construct instead of `PyObject_Init` and friends? It seems the segfault (on gcc4.9 only) in the previous commit was due to having a `Read` member in `khmer_ReadPairIterator_Object` instead of a `Read *`.

This version also runs in about ~53s. Next up investigating why this doesn't run faster and maybe a small toy example to fully understand how memory gets initialised for python instances :-/ My guess as to why this isn't faster would be that we switch from C to python to C to python. Instead of handing a big chunk of work off to C land (`numpy` style).