@ctb 2016-11-22
Compare with @camillescott's Cythonized `broken_paired_reader`:

```
In [5]: %timeit [record for record in BrokenPairedReader(FastxParser('test-mixed.fa'))]
100 loops, best of 3: 3.49 ms per loop

In [6]: %timeit [record for record in broken_paired_reader(screed.open('test-mixed.fa'))]
100 loops, best of 3: 17.8 ms per loop
```