@betatim 2016-11-22
Exactly, it would suggest that there is something obviously wrong with my c++ 😢