@betatim 2016-11-23
With dc0d0a6, modified `setup.py` (to get it compile), and modified `extract-paired-reads.py` like so:
```python
from khmer._oxli import streaming
screed_iter = streaming.FastxParser(infile)
for index, is_pair, read1, read2 in streaming.BrokenPairedReader(screed_iter):
    ...
```
the above bench on ecoli takes ~60s. Conclusion: we can make the broken paired read parsing faster (yay!) but that isn't where we are spending our time (at least in this application) (:cry:)