@ctb 2016-12-20
what is your thinking on this now?