@betatim 2016-12-21
Still think that reading stuff from disk doesn't seem to be the bottle neck. However both Camille's cython and this PR are quicker than the python version, so let's merge it.