@betatim 2016-12-22
New machine, new benchmarks (`time scripts/extract-paired-reads.py ecoli_ref-5m.fastq.gz -s /dev/null -p /dev/null`).

* `master`: 31s
* this branch: 23s

This is pretty big! I am worried about correctness but all the tests pass 😕 The code is quite messy as it is, I will tidy it up a bit and think through that gain more confidence in it 🤞.