@betatim 2017-01-25
I thought it would be a good idea to get this done before #1591 but somehow nearly ended in disaster. Tried to replace the internals of `broken_paired_reader` so that they can take reads from `ReadParser(fname).iter_...` and `screed.open(fname)`. New attempt.