@betatim 2017-04-07
Thoughts on punting this till after 2.0? The code proposed in this PR is pretty messy and ergh, post 2.0 there is the ope of cython where this work has already been done. So I would close this PR and reevaluate after 2.0