@ctb 2017-04-07
+1

On Fri, Apr 07, 2017 at 02:23:05AM -0700, Tim Head wrote:
> Thoughts on punting this till after 2.0? The code proposed in this PR is pretty messy and ergh, post 2.0 there is the ope of cython where this work has already been done. So I would close this PR and reevaluate after 2.0
> 
> -- 
> You are receiving this because you commented.
> Reply to this email directly or view it on GitHub:
> https://github.com/dib-lab/khmer/pull/1538#issuecomment-292487016
-- 
C. Titus Brown, ctbrown@ucdavis.edu