@betatim 2017-04-07
Closed in favour of cython (or a new attempt).

---

Please don't delete the branch so we can recover the code from this PR.